feat: formalising multistart
